+++ app/components/Navbar.tsx
@@ -15,12 +15,12 @@ const Navbar = () => {
       <input id="my-drawer-3" type="checkbox" className="drawer-toggle" />
       <div className="drawer-content flex flex-col">
         {/* Navbar */}
-        <div className="navbar bg-gradient-to-r from-blue-600 via-purple-600 to-indigo-600 text-white shadow-lg w-full">
+        <div className="navbar backdrop-blur-md bg-black/30 border-b border-gray-700/50 text-white shadow-lg w-full fixed top-0 z-50">
           <div className="flex-none lg:hidden">
             <label
               htmlFor="my-drawer-3"
               aria-label="open sidebar"
-              className="btn btn-square btn-ghost text-white"
+              className="btn btn-square btn-ghost text-white hover:bg-orange-500/20"
             >
               <svg
                 xmlns="http://www.w3.org/2000/svg"
@@ -37,25 +37,39 @@ const Navbar = () => {
               </svg>
             </label>
           </div>
-          <div className="mx-2 flex-1 px-2 text-2xl font-bold">Nexarb</div>
+          
+          <div className="mx-2 flex-1 px-2">
+            <Link href="/" className="flex items-center gap-2">
+              <span className="text-2xl font-bold bg-gradient-to-r from-orange-400 to-pink-500 bg-clip-text text-transparent">
+                Nexarb
+              </span>
+            </Link>
+          </div>
+
           <div className="hidden flex-none lg:block">
-            <ul className="menu menu-horizontal flex items-center px-4 py-2">
+            <ul className="menu menu-horizontal flex items-center px-4 py-2 gap-4">
               {path === "/" && (
-                <li className="mr-4 flex items-center">
-                  <Link href="/app" className={buttonClass}>
+                <li className="flex items-center">
+                  <Link
+                    href="/app"
+                    className="px-6 py-2 bg-gradient-to-r from-orange-400 to-pink-500 text-white font-semibold rounded-full shadow-[0_0_15px_rgba(251,146,60,0.5)] hover:shadow-[0_0_20px_rgba(251,146,60,0.7)] transition-all duration-300"
+                  >
                     Launch App
                   </Link>
                 </li>
               )}
               {path !== "/" && (
-                <li className="mr-4 flex items-center">
+                <li className="flex items-center">
                   <WalletButton />
                 </li>
               )}
               {path !== '/' && path !== '/app' && (
-                <li className="mr-4 flex items-center">
-                  <button onClick={() => router.push('/app')} className={buttonClass}>
-                    Go Back to Configuration
+                <li className="flex items-center">
+                  <button
+                    onClick={() => router.push('/app')}
+                    className="px-6 py-2 bg-gradient-to-r from-gray-800 to-gray-900 text-white font-semibold rounded-full border border-gray-700 hover:border-orange-500/50 transition-all duration-300"
+                  >
+                    Go Back to App
                   </button>
                 </li>
               )}
@@ -63,6 +77,8 @@ const Navbar = () => {
           </div>
         </div>
       </div>
+      
+      {/* Sidebar */}
       {isConnected && (
         <div className="drawer-side">
           <label
@@ -70,38 +86,42 @@ const Navbar = () => {
             aria-label="close sidebar"
             className="drawer-overlay"
           ></label>
-          <ul className="menu bg-base-200 min-h-full w-80 p-4">
-            {path == "/" && (
-              <>
-                <li className="mx-6">
-                  <Link href="/voice" className={buttonClass}>
+          <div className="menu min-h-full w-80 bg-gradient-to-b from-gray-900 to-black border-r border-gray-800/50 p-4">
+            <div className="flex flex-col gap-4">
+              {path === "/" && (
+                <>
+                  <Link
+                    href="/voice"
+                    className="px-6 py-3 bg-gradient-to-r from-orange-400 to-pink-500 text-white font-semibold rounded-xl shadow-lg hover:shadow-orange-500/20 transition duration-300 text-center"
+                  >
                     Launch App with Voice
                   </Link>
-                </li>
-                <li>
-                  <Link href="/chat" className={buttonClass}>
+                  <Link
+                    href="/chat"
+                    className="px-6 py-3 bg-gradient-to-r from-gray-800 to-gray-900 text-white font-semibold rounded-xl border border-gray-700 hover:border-orange-500/50 transition duration-300 text-center"
+                  >
                     Launch App with Chat
                   </Link>
-                </li>
-              </>
-            )}
-            {path == "/chat" && (
-              <Link
-                href="/voice"
-                className="px-4 py-2 bg-white text-blue-600 font-semibold rounded-full shadow-md hover:bg-gray-200 transition duration-300"
-              >
-                Switch to Voice
-              </Link>
-            )}
-            {path == "/voice" && (
-              <Link
-                href="/chat"
-                className="px-4 py-2 bg-white text-blue-600 font-semibold rounded-full shadow-md hover:bg-gray-200 transition duration-300"
-              >
-                Switch to Chat
-              </Link>
-            )}
-          </ul>
+                </>
+              )}
+              {path === "/chat" && (
+                <Link
+                  href="/voice"
+                  className="px-6 py-3 bg-gradient-to-r from-orange-400 to-pink-500 text-white font-semibold rounded-xl shadow-lg hover:shadow-orange-500/20 transition duration-300 text-center"
+                >
+                  Switch to Voice
+                </Link>
+              )}
+              {path === "/voice" && (
+                <Link
+                  href="/chat"
+                  className="px-6 py-3 bg-gradient-to-r from-orange-400 to-pink-500 text-white font-semibold rounded-xl shadow-lg hover:shadow-orange-500/20 transition duration-300 text-center"
+                >
+                  Switch to Chat
+                </Link>
+              )}
+            </div>
+          </div>
         </div>
       )}
     </div>
