+++ app/globals.css
@@ -16,8 +16,13 @@
 }
 
 body {
-  color: var(--foreground);
-  background: var(--background);
+  background: black;
+  color: white;
   font-family: Arial, Helvetica, sans-serif;
   min-height: 100vh;
+}
+
+/* Add this new utility class */
+.page-with-navbar {
+  padding-top: 5rem; /* 80px */
 }
