+++ prisma/schema.prisma
@@ -33,6 +33,14 @@ model SpecialUserCodes {
   used_by         String
 }
 
+model ArbitrumWallets {
+  id              Int      @id @default(autoincrement())
+  wallet_address  String   @unique
+  private_key     String
+  is_verified     Boolean @default(false)
+  mnemonic        String?
+}
+
 model BaseWallets {
   id              Int      @id @default(autoincrement())
   wallet_address  String   @unique
@@ -40,6 +48,7 @@ model BaseWallets {
   is_verified     Boolean @default(false)
   mnemonic        String?
 }
+
 
 model Admins {
   wallet_address String @id
